rename past session dto for reuse as all responses and request body use that structure now
